// src/types/wine.ts
export interface WineData {
    Class: any;
    Alcohol: number;
    "Malic Acid": number;
    Ash: number;
    "Alcalinity of ash": number;
    Magnesium: number;
    "Total phenols": number;
    Flavanoids: number;
    "Nonflavanoid phenols": number;
    Proanthocyanins: string;
    "Color intensity": number;
    Hue: number;
    "OD280/OD315 of diluted wines": string;
    Unknown: number;
}